Move testing helper function to common place
We'll need to reuse this functionality later for multi-dimensional
quadrature tests, so let's put it in a common place
